+++ app/code/Magento/Catalog/Model/Product/Link/SaveHandler.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright © 2016 Magento. All rights reserved.
+ * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
 
@@ -31,6 +31,7 @@ class SaveHandler
     private $linkResource;
 
     /**
+     * SaveHandler constructor.
      * @param MetadataPool $metadataPool
      * @param Link $linkResource
      * @param ProductLinkRepositoryInterface $productLinkRepository
@@ -40,7 +41,6 @@ class SaveHandler
         Link $linkResource,
         ProductLinkRepositoryInterface $productLinkRepository
     ) {
-
         $this->metadataPool = $metadataPool;
         $this->linkResource = $linkResource;
         $this->productLinkRepository = $productLinkRepository;
@@ -54,13 +54,52 @@ class SaveHandler
      */
     public function execute($entityType, $entity)
     {
-        /** @var \Magento\Catalog\Api\Data\ProductInterface $entity*/
-        foreach ($this->productLinkRepository->getList($entity) as $link) {
-            $this->productLinkRepository->delete($link);
+        $link = $entity->getData($this->metadataPool->getMetadata($entityType)->getLinkField());
+        if ($this->linkResource->hasProductLinks($link)) {
+            /** @var \Magento\Catalog\Api\Data\ProductInterface $entity */
+            foreach ($this->productLinkRepository->getList($entity) as $link) {
+                $this->productLinkRepository->delete($link);
+            }
         }
+
+        // Build links per type
+        $linksByType = [];
         foreach ($entity->getProductLinks() as $link) {
-            $this->productLinkRepository->save($link);
+            $linksByType[$link->getLinkType()][] = $link;
+        }
+
+        // Set array position as a fallback position if necessary
+        foreach ($linksByType as $linkType => $links) {
+            if (!$this->hasPosition($links)) {
+                array_walk($linksByType[$linkType], function ($productLink, $position) {
+                    $productLink->setPosition(++$position);
+                });
+            }
+        }
+
+        // Flatten multi-dimensional linksByType in ProductLinks
+        $productLinks = array_reduce($linksByType, 'array_merge', []);
+
+        if (count($productLinks) > 0) {
+            foreach ($entity->getProductLinks() as $link) {
+                $this->productLinkRepository->save($link);
+            }
         }
         return $entity;
+    }
+
+    /**
+     * Check if at least one link without position
+     * @param array $links
+     * @return bool
+     */
+    private function hasPosition(array $links)
+    {
+        foreach ($links as $link) {
+            if (!array_key_exists('position', $link->getData())) {
+                return false;
+            }
+        }
+        return true;
     }
 }
